Fix phantom -0.5% loss + always read news + harden stale-trade detection
Critical fixes:
- broker_tv_paper: paper fills now use actual market price, not the
  inflated limit price. Previously every position showed an immediate
  ~0.5% loss because get_positions reads market while submit_buy stored
  the +ATR-offset limit price as avg_cost.
- sentiment: keyword-based fallback so news headlines are actually
  ANALYSED when GROQ_API_KEY is missing or LLM fails (previously
  defaulted to neutral 5 even with available headlines).
- heartbeat: stale-trade auto-close now requires 3 consecutive misses
  to defend against transient broker API hiccups silently closing all
  open trades.

Smaller cleanups:
- news_service: word-boundary regex for ticker matching (no more 'ON'
  matching every headline containing 'on').
- heartbeat: NameError in daily summary log line (today_trades undef).
- telegram_bot: drop unused buys_line var.
- broker.py / budget.py: qty type annotations -> float for fractional.

+++ broker.py
@@ -167,11 +167,11 @@ def get_position(ticker: str) -> dict | None:
     return _get_broker().get_position(ticker)
 
 
-def submit_buy(ticker: str, qty: int, price: float | None = None) -> dict:
+def submit_buy(ticker: str, qty: float, price: float | None = None) -> dict:
     return _get_broker().submit_buy(ticker, qty, price)
 
 
-def submit_sell(ticker: str, qty: int | None = None) -> dict:
+def submit_sell(ticker: str, qty: float | None = None) -> dict:
     return _get_broker().submit_sell(ticker, qty)
